Refactored PartialEclipseTracker into several independent classes.
* EclipseTracker: base class for tracking eclipse images using optimization

* PartialEclipseTracker: class derived from EclipseTracker, which is
    specialized for partial eclipse images.

* PartialEclipsePreprocessor: preprocesses partial eclipse images for
   tracking via PartialEclipseTracker.

This allows code reuse for tracking total eclipse images.
