+++ processing/run_track_partials.py
@@ -3,6 +3,7 @@ import copy
 import matplotlib.pyplot as plt
 import numpy as np
 
+import eclipse_tracker
 import partial_eclipse_tracker
 
 
@@ -17,13 +18,16 @@ def main(plot_only: bool = False):
       guess_moon_dc_dt=0.7 * 4 * 130 / 9000,
   )
   tracker = partial_eclipse_tracker.PartialEclipseTracker(options)
-  tracker.load_preprocess_from_file('preprocess_partials')
+
+  preprocessor = partial_eclipse_tracker.PartialEclipsePreprocessor()
+  preprocessor.load_from_file('preprocess_partials')
+  preprocessor.init_partial_eclipse_tracker(tracker)
 
   if plot_only:
-    track = partial_eclipse_tracker.load_track('partial_track')
+    track = eclipse_tracker.load_track('partial_track')
   else:
     track = tracker.track_sun_and_moon()
-    partial_eclipse_tracker.save_track('partial_track', track)
+    eclipse_tracker.save_track('partial_track', track)
 
   # Print tracking results.
   print_track = copy.copy(track)
@@ -34,7 +38,7 @@ def main(plot_only: bool = False):
 
   # Plot final aligned images.
   ind_to_plot = np.asarray(
-      np.linspace(0, tracker.is_sun.shape[0] - 1, 30),
+      np.linspace(0, tracker.bw_images.shape[0] - 1, 30),
       dtype=int
   )
   tracker.plot_track(track, ind_to_plot)
